pyproject.toml
[project]
name = "immich-adapter"
version = "0.1.0"
description = "FastAPI server that adapts the Immich API to the Gumnut API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "b64uuid>=0.1",
    "fastapi[standard]>=0.116.1",
    "gumnut-sdk",
    "pydantic-settings>=2.10.1",
    "pytest>=8.4.2",
    "python-socketio>=5.13.0",
    "sentry-sdk[fastapi]>=2.37.1",
    "shortuuid>=1.0.13",
]

[dependency-groups]
dev = [
    "pyright>=1.1.403",
    "ruff>=0.9.2",
]


[tool.uv.sources]
gumnut-sdk = { git = "https://github.com/gumnut-ai/photos-sdk-python.git" }
